+++ spec/outputs/kusto/ingestor_spec.rb
@@ -18,6 +18,7 @@ describe LogStash::Outputs::Kusto::Ingestor do
   let(:proxy_protocol) { "http" }
   let(:json_mapping) { "mymapping" }
   let(:delete_local) { false }
+  let(:dynamic_routing) { false }
   let(:logger) { spy('logger') }
 
   describe '#initialize' do
@@ -26,7 +27,7 @@ describe LogStash::Outputs::Kusto::Ingestor do
       # note that this will cause an internal error since connection is being tried.
       # however we still want to test that all the java stuff is working as expected
       expect { 
-        ingestor = described_class.new(ingest_url, app_id, app_key, app_tenant, managed_identity, cliauth, database, table, json_mapping, delete_local, proxy_host, proxy_port,proxy_protocol, logger)
+        ingestor = described_class.new(ingest_url, app_id, app_key, app_tenant, managed_identity, cliauth, database, table, json_mapping, dynamic_routing, delete_local, proxy_host, proxy_port,proxy_protocol, logger)
         ingestor.stop
       }.not_to raise_error
     end
@@ -37,7 +38,7 @@ describe LogStash::Outputs::Kusto::Ingestor do
       dynamic_name_array.each do |test_database|
         it "with database: #{test_database}" do
           expect {
-            ingestor = described_class.new(ingest_url, app_id, app_key, app_tenant, managed_identity, cliauth, test_database, table, json_mapping, delete_local, proxy_host, proxy_port,proxy_protocol,logger)
+            ingestor = described_class.new(ingest_url, app_id, app_key, app_tenant, managed_identity, cliauth, test_database, table, json_mapping, dynamic_routing, delete_local, proxy_host, proxy_port,proxy_protocol,logger)
             ingestor.stop
           }.to raise_error(LogStash::ConfigurationError)          
         end
@@ -48,7 +49,7 @@ describe LogStash::Outputs::Kusto::Ingestor do
       dynamic_name_array.each do |test_table|
         it "with database: #{test_table}" do
           expect {
-            ingestor = described_class.new(ingest_url, app_id, app_key, app_tenant, managed_identity, cliauth, database, test_table, json_mapping, delete_local, proxy_host, proxy_port,proxy_protocol,logger)
+            ingestor = described_class.new(ingest_url, app_id, app_key, app_tenant, managed_identity, cliauth, database, test_table, json_mapping, dynamic_routing, delete_local, proxy_host, proxy_port,proxy_protocol,logger)
             ingestor.stop
           }.to raise_error(LogStash::ConfigurationError)          
         end
@@ -59,9 +60,20 @@ describe LogStash::Outputs::Kusto::Ingestor do
       dynamic_name_array.each do |json_mapping|
         it "with database: #{json_mapping}" do
           expect {
-            ingestor = described_class.new(ingest_url, app_id, app_key, app_tenant, managed_identity, cliauth, database, table, json_mapping, delete_local, proxy_host, proxy_port,proxy_protocol,logger)
+            ingestor = described_class.new(ingest_url, app_id, app_key, app_tenant, managed_identity, cliauth, database, table, json_mapping, dynamic_routing, delete_local, proxy_host, proxy_port,proxy_protocol,logger)
             ingestor.stop
           }.to raise_error(LogStash::ConfigurationError)          
+        end
+      end
+    end
+
+    context 'allows dynamic database/table/mapping when dynamic routing is enabled' do
+      dynamic_name_array.each do |dynamic_value|
+        it "with dynamic value: #{dynamic_value}" do
+          expect {
+            ingestor = described_class.new(ingest_url, app_id, app_key, app_tenant, managed_identity, cliauth, dynamic_value, dynamic_value, dynamic_value, true, delete_local, proxy_host, proxy_port, proxy_protocol, logger)
+            ingestor.stop
+          }.not_to raise_error
         end
       end
     end
@@ -69,7 +81,7 @@ describe LogStash::Outputs::Kusto::Ingestor do
     context 'proxy protocol has to be http or https' do
       it "with proxy protocol: socks" do
         expect {
-          ingestor = described_class.new(ingest_url, app_id, app_key, app_tenant, managed_identity, cliauth, database, table, json_mapping, delete_local, proxy_host, proxy_port,'socks',logger)
+          ingestor = described_class.new(ingest_url, app_id, app_key, app_tenant, managed_identity, cliauth, database, table, json_mapping, dynamic_routing, delete_local, proxy_host, proxy_port,'socks',logger)
           ingestor.stop
         }.to raise_error(LogStash::ConfigurationError)          
       end
@@ -78,12 +90,49 @@ describe LogStash::Outputs::Kusto::Ingestor do
     context 'one of appid or managedid has to be provided' do
       it "with empty managed identity and appid" do
         expect {
-          ingestor = described_class.new(ingest_url, "", app_key, app_tenant, "", cliauth, database, table, json_mapping, delete_local, proxy_host, proxy_port,'socks',logger)
+          ingestor = described_class.new(ingest_url, "", app_key, app_tenant, "", cliauth, database, table, json_mapping, dynamic_routing, delete_local, proxy_host, proxy_port,'socks',logger)
           ingestor.stop
         }.to raise_error(LogStash::ConfigurationError)          
       end
     end
 
+  end
+
+  describe '#decode_routing_target' do
+    let(:ingestor) do
+      described_class.new(ingest_url, app_id, app_key, app_tenant, managed_identity, cliauth, "%{db}", "%{table}", "%{mapping}", true, delete_local, proxy_host, proxy_port, proxy_protocol, logger)
+    end
+
+    after(:each) { ingestor.stop }
+
+    it 'decodes database, table and mapping from the file name' do
+      target = ingestor.decode_routing_target("/tmp/kusto/2024-01-01.kusto~mydb~mytable~mymapping")
+      expect(target[:database]).to eq('mydb')
+      expect(target[:table]).to eq('mytable')
+      expect(target[:mapping]).to eq('mymapping')
+    end
+
+    it 'treats an empty mapping segment as no mapping' do
+      target = ingestor.decode_routing_target("/tmp/kusto/2024-01-01.kusto~mydb~mytable~")
+      expect(target[:database]).to eq('mydb')
+      expect(target[:table]).to eq('mytable')
+      expect(target[:mapping]).to be_nil
+    end
+
+    it 'is unaffected by dots in the path prefix' do
+      target = ingestor.decode_routing_target("/tmp/kusto/2024.01.01-10.30.kusto~mydb~mytable~mymapping")
+      expect(target[:database]).to eq('mydb')
+      expect(target[:table]).to eq('mytable')
+      expect(target[:mapping]).to eq('mymapping')
+    end
+
+    it 'returns nil when the routing marker is absent (dead-letter file)' do
+      expect(ingestor.decode_routing_target("/tmp/kusto/_filepath_failures")).to be_nil
+    end
+
+    it 'returns nil when database or table did not resolve' do
+      expect(ingestor.decode_routing_target("/tmp/kusto/2024-01-01.kusto~mydb~")).to be_nil
+    end
   end
 
   # describe 'receiving events' do